requirements.txt
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
flask==3.0.3
fsspec==2024.6.1
huggingface-hub==0.24.3
idna==3.7
importlib-metadata==8.2.0
ipython==8.18.1
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-core==5.7.2
jupyterlab-widgets==3.0.11
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
open3d==0.18.0
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
prompt-toolkit==3.0.47
pure-eval==0.2.3
pygments==2.18.0
pywin32==306
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rpds-py==0.19.1
safetensors==0.4.3
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
timm==1.0.8
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
traitlets==5.14.3
typing-extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
werkzeug==3.0.3
widgetsnbextension==4.0.11
zipp==3.19.2
